Extract shared manifest utilities for artifact checks
- Move manifest parsing and target collection into `scripts/manifest-utils.ts`
- Tighten package path normalization to reject invalid artifact paths
- Reuse the shared helpers in artifact verification scripts
