+++ scripts/check-publish-artifacts.ts
@@ -6,6 +6,12 @@ import {
 	checkPackedAgentDocs,
 	supportedAgentDocsPackages,
 } from './agent-docs/check-budgets';
+import type { PackageManifest } from './manifest-utils';
+import {
+	collectManifestTargets,
+	readManifest,
+	wildcardToRegExp,
+} from './manifest-utils';
 
 type PackedFile = {
 	path: string;
@@ -16,21 +22,6 @@ type PackResult = {
 	name: string;
 	version: string;
 	files: PackedFile[];
-};
-
-type PackageManifest = {
-	name?: string;
-	private?: boolean;
-	main?: string;
-	module?: string;
-	types?: string;
-	bin?: string | Record<string, string>;
-	exports?: unknown;
-};
-
-type ManifestTarget = {
-	source: string;
-	target: string;
 };
 
 const ROOT = process.cwd();
@@ -101,86 +92,6 @@ const rootTw3ProxyContents: Record<string, string> = {
 	'styles.tw3.css': '@import "./dist/styles.tw3.css";',
 	'iab/styles.tw3.css': '@import "../dist/iab/styles.tw3.css";',
 };
-
-function readManifest(packageDir: string): PackageManifest {
-	return JSON.parse(
-		readFileSync(join(packageDir, 'package.json'), 'utf8')
-	) as PackageManifest;
-}
-
-function normalizePackagePath(target: string): string | null {
-	if (target.startsWith('/') || target.includes('://')) {
-		return null;
-	}
-
-	return target.replace(/^\.\//, '').replaceAll('\\', '/');
-}
-
-function collectExportTargets(
-	value: unknown,
-	targets: ManifestTarget[],
-	source: string
-) {
-	if (typeof value === 'string') {
-		const normalized = normalizePackagePath(value);
-		if (normalized) {
-			targets.push({ source, target: normalized });
-		}
-		return;
-	}
-
-	if (Array.isArray(value)) {
-		for (const item of value) {
-			collectExportTargets(item, targets, source);
-		}
-		return;
-	}
-
-	if (value && typeof value === 'object') {
-		for (const [key, item] of Object.entries(value)) {
-			collectExportTargets(item, targets, `${source}[${JSON.stringify(key)}]`);
-		}
-	}
-}
-
-function collectManifestTargets(manifest: PackageManifest): ManifestTarget[] {
-	const targets: ManifestTarget[] = [];
-
-	for (const key of ['main', 'module', 'types'] as const) {
-		const value = manifest[key];
-		if (typeof value !== 'string') {
-			continue;
-		}
-
-		const normalized = normalizePackagePath(value);
-		if (normalized) {
-			targets.push({ source: key, target: normalized });
-		}
-	}
-
-	if (typeof manifest.bin === 'string') {
-		const normalized = normalizePackagePath(manifest.bin);
-		if (normalized) {
-			targets.push({ source: 'bin', target: normalized });
-		}
-	} else if (manifest.bin && typeof manifest.bin === 'object') {
-		for (const [name, value] of Object.entries(manifest.bin)) {
-			const normalized = normalizePackagePath(value);
-			if (normalized) {
-				targets.push({ source: `bin.${name}`, target: normalized });
-			}
-		}
-	}
-
-	collectExportTargets(manifest.exports, targets, 'exports');
-
-	return targets;
-}
-
-function wildcardToRegExp(pattern: string): RegExp {
-	const escaped = pattern.replace(/[.+?^${}()|[\]\\]/g, '\\$&');
-	return new RegExp(`^${escaped.replaceAll('*', '.+')}$`);
-}
 
 function scanPackedManifestTargets(
 	manifest: PackageManifest,
